% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_methods-dbproj_char-tuple.R
\name{alter.dbproj_char_tuple}
\alias{alter.dbproj_char_tuple}
\title{S3 Method - Alter Internal State of S3 Object Instances}
\usage{
\method{alter}{dbproj_char_tuple}(obj, ...)
}
\arguments{
\item{obj}{S3 Object}

\item{...}{R ellipsis}
}
\description{
S3 Method - Alter Internal State of S3 Object Instances
}
\examples{
\dontrun{
alter(obj, ...)
}
}
