% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_generics.R
\name{alter}
\alias{alter}
\title{S3 Generic - Alter Internal State of S3 Object Instances}
\usage{
alter(obj, ...)
}
\arguments{
\item{obj}{S3 Object}

\item{...}{R ellipsis}
}
\value{
S3 Object
}
\description{
S3 Generic - Alter Internal State of S3 Object Instances
}
\examples{
\dontrun{
alter(s3_obj)
}
}
