% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_class-dbproj_char-tuple.R
\name{char_tuple}
\alias{char_tuple}
\title{S3 Helper Function for Class 'char_tuple'}
\usage{
char_tuple(x)
}
\arguments{
\item{x}{character}
}
\value{
S3 Object
}
\description{
S3 Helper Function for Class 'char_tuple'
}
\examples{
\dontrun{
output <- char_tuple(c('a', 'b', 'c'))
}
}
