% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_class-dbproj_column.R
\name{column}
\alias{column}
\title{S3 Helper Function for Class 'column'}
\usage{
column(name)
}
\arguments{
\item{name}{data.frame}
}
\value{
S3 Object
}
\description{
S3 Helper Function for Class 'column'
}
\examples{
\dontrun{
test <- column(IMF_DATA)
}
}
