% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_methods-db_schema.R
\name{contents.db_schema}
\alias{contents.db_schema}
\title{S3 Method - CRUD 'Read' Operation}
\usage{
\method{contents}{db_schema}(obj, ...)
}
\arguments{
\item{obj}{S3 Object}

\item{...}{R ellipsis}
}
\description{
S3 Method - CRUD 'Read' Operation
}
\examples{
\dontrun{
result <- contents(obj, ...)
}
}
