% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_methods-dbproj_db.R
\name{contents.dbproj_db}
\alias{contents.dbproj_db}
\title{S3 Method - CRUD 'Read' Operation}
\usage{
\method{contents}{dbproj_db}(obj, ...)
}
\arguments{
\item{obj}{S3 Object}

\item{...}{R ellipsis}
}
\description{
S3 Method - CRUD 'Read' Operation
}
\examples{
\dontrun{
result <- contents(obj, ...)
}
}
