% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_generics.R
\name{contents}
\alias{contents}
\title{S3 Generic - CRUD 'Read' Operation}
\usage{
contents(obj, ...)
}
\arguments{
\item{obj}{S3 Object}

\item{...}{R ellipsis}
}
\description{
S3 Generic - CRUD 'Read' Operation
}
\examples{
\dontrun{
result <- contents(obj, ...)
}
}
