% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_class-dbproj_db.R
\name{db}
\alias{db}
\title{S3 Helper Function for Class 'dbproj_db'}
\usage{
db(name, schema)
}
\arguments{
\item{name}{character}

\item{schema}{list}
}
\value{
S3 Object
}
\description{
S3 Helper Function for Class 'dbproj_db'
}
\examples{
\dontrun{
test <- db(name)
}
}
