% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_class-db_database.R
\name{db_database}
\alias{db_database}
\title{S3 Helper Function for Class 'db_database'}
\usage{
db_database(name, schemas)
}
\arguments{
\item{name}{character}

\item{schemas}{list}
}
\value{
S3 Object
}
\description{
S3 Helper Function for Class 'db_database'
}
\examples{
\dontrun{
test <- db_database(name)
}
}
