% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_proj_template.R
\name{db_proj_template}
\alias{db_proj_template}
\title{S3 Class Helper Function for 'db_proj_template'}
\usage{
db_proj_template(type)
}
\arguments{
\item{type}{character - Specify database type for template}
}
\value{
S3 Object
}
\description{
S3 Class Helper Function for 'db_proj_template'
}
\examples{
\dontrun{
db_proj_template('mssql')
}
}
