% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{db_proj_template_types}
\alias{db_proj_template_types}
\title{List of Valid Values for 'db_proj_template' Setup Types}
\description{
List of Valid Values for 'db_proj_template' Setup Types
}
\keyword{data}
