% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_class-db_schema.R
\name{db_schema}
\alias{db_schema}
\title{S3 Helper Function for Class 'db_schema'}
\usage{
db_schema(x)
}
\arguments{
\item{x}{data.frame}
}
\value{
S3 Object
}
\description{
S3 Helper Function for Class 'db_schema'
}
\examples{
\dontrun{
test <- db_schema(IMF_DATA)
}
}
