% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_methods-db_database.R
\name{delete.db_database}
\alias{delete.db_database}
\title{S3 Method - CRUD 'Delete' Operation}
\usage{
\method{delete}{db_database}(obj, ...)
}
\arguments{
\item{obj}{S3 Object}

\item{...}{R ellipsis}
}
\description{
S3 Method - CRUD 'Delete' Operation
}
\examples{
\dontrun{
delete(obj, ...)
}
}
