% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_methods-dbproj_char-tuple.R
\name{delete.dbproj_char_tuple}
\alias{delete.dbproj_char_tuple}
\title{S3 Method - CRUD 'Delete' Operation}
\usage{
\method{delete}{dbproj_char_tuple}(obj, x, in_place = FALSE, ...)
}
\arguments{
\item{obj}{S3 Object}

\item{x}{character}

\item{in_place}{logical}

\item{...}{R ellipsis}
}
\description{
S3 Method - CRUD 'Delete' Operation
}
\examples{
\dontrun{
delete(obj, x)
output <- delete(obj, x, in_place = FALSE)
}
}
