% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_methods-dbproj_schema.R
\name{delete.dbproj_schema}
\alias{delete.dbproj_schema}
\title{S3 Method - CRUD 'Delete' Operation}
\usage{
\method{delete}{dbproj_schema}(obj, ...)
}
\arguments{
\item{obj}{S3 Object}

\item{...}{R ellipsis}
}
\description{
S3 Method - CRUD 'Delete' Operation
}
\examples{
\dontrun{
delete(obj, ...)
}
}
