% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_generics.R
\name{delete}
\alias{delete}
\title{S3 Generic - CRUD 'Delete' Operation}
\usage{
delete(obj, ...)
}
\arguments{
\item{obj}{S3 Object}

\item{...}{R ellipsis}
}
\description{
S3 Generic - CRUD 'Delete' Operation
}
\examples{
\dontrun{
delete(obj, ...)
}
}
