% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_methods-dbproj_column.R
\name{equals.dbproj_column}
\alias{equals.dbproj_column}
\title{S3 Method - Test Equality of S3 'dbproj_column' Instances}
\usage{
\method{equals}{dbproj_column}(obj, ...)
}
\arguments{
\item{obj}{S3 Object}

\item{...}{R ellipsis}
}
\value{
S3 Object
}
\description{
S3 Method - Test Equality of S3 'dbproj_column' Instances
}
\examples{
\dontrun{
test <- equals(s3_obj, FALSE)
}
}
