% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_methods-dbproj_db.R
\name{equals.dbproj_db}
\alias{equals.dbproj_db}
\title{S3 Method - Test Equality of S3 'dbproj_db' Instances}
\usage{
\method{equals}{dbproj_db}(obj, ...)
}
\arguments{
\item{obj}{S3 Object}

\item{...}{R ellipsis}
}
\value{
S3 Object
}
\description{
S3 Method - Test Equality of S3 'dbproj_db' Instances
}
\examples{
\dontrun{
test <- equals(s3_obj, FALSE)
}
}
