% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_generics.R
\name{equals}
\alias{equals}
\title{S3 Generic - Test Equality of S3 Object Instances}
\usage{
equals(obj, ..., bool)
}
\arguments{
\item{obj}{S3 Object}

\item{...}{R ellipsis}
}
\value{
S3 Object
}
\description{
S3 Generic - Test Equality of S3 Object Instances
}
\examples{
\dontrun{
test <- equals(s3_obj, FALSE)
}
}
