% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-gen_test_data.R
\name{gen_test_char}
\alias{gen_test_char}
\title{Generate Dummy Character Vectors for Package Testing}
\usage{
gen_test_char(n = 10, min = 5, max = 20)
}
\arguments{
\item{n}{numeric}

\item{min}{numeric}

\item{max}{numeric}
}
\value{
character
}
\description{
Generate Dummy Character Vectors for Package Testing
}
\examples{
\dontrun{
output <- gen_test_char(n = 20, min = 5, max = 20)
}
}
