% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-gen_test_data.R
\name{gen_test_numeric}
\alias{gen_test_numeric}
\title{Generate Dummy Numeric Vectors for Package Testing}
\usage{
gen_test_numeric(n = 10, min = 2, max = 100, as_int = FALSE)
}
\arguments{
\item{n}{numeric}

\item{min}{numeric}

\item{max}{numeric}

\item{as_int}{logical}
}
\value{
character
}
\description{
Generate Dummy Numeric Vectors for Package Testing
}
\examples{
\dontrun{
output <- gen_test_numeric(n = 20, min = 5, max = 19)
}
}
