% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_generics.R
\name{generate}
\alias{generate}
\title{S3 Generic - Generate Dummy Test Data}
\usage{
generate(obj, ...)
}
\arguments{
\item{obj}{S3 Object}

\item{...}{R ellipsis}
}
\description{
S3 Generic - Generate Dummy Test Data
}
\examples{
\dontrun{
generate(obj, ...)
}
}
