% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_methods-db_database.R
\name{get.db_database}
\alias{get.db_database}
\title{S3 Method - CRUD 'Read (Get)' Operation}
\usage{
get.db_database(obj, ...)
}
\arguments{
\item{obj}{S3 Object}

\item{...}{R ellipsis}
}
\description{
S3 Method - CRUD 'Read (Get)' Operation
}
\examples{
\dontrun{
result <- get(obj, ...)
}
}
