% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_methods-db_database.R
\name{insert.db_database}
\alias{insert.db_database}
\title{S3 Method - CRUD 'Insert' Operation}
\usage{
\method{insert}{db_database}(obj, ...)
}
\arguments{
\item{obj}{S3 Object}

\item{...}{R ellipsis}
}
\description{
S3 Method - CRUD 'Insert' Operation
}
\examples{
\dontrun{
insert(obj, ...)
}
}
