% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_methods-dbproj_char-tuple.R
\name{insert.dbproj_char_tuple}
\alias{insert.dbproj_char_tuple}
\title{S3 Method - CRUD 'Insert' Operation}
\usage{
\method{insert}{dbproj_char_tuple}(obj, ...)
}
\arguments{
\item{obj}{S3 Object}

\item{...}{R ellipsis}
}
\description{
S3 Method - CRUD 'Insert' Operation
}
\examples{
\dontrun{
insert(obj, ...)
}
}
