% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_generics.R
\name{insert}
\alias{insert}
\title{S3 Generic - CRUD 'Insert' Operation}
\usage{
insert(obj, ...)
}
\arguments{
\item{obj}{S3 Object}

\item{...}{R ellipsis}
}
\description{
S3 Generic - CRUD 'Insert' Operation
}
\examples{
\dontrun{
insert(obj, ...)
}
}
