% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_class-db_database.R
\name{new_db_database}
\alias{new_db_database}
\title{S3 Constructor for Class 'db_database'}
\usage{
new_db_database(name, schemas)
}
\arguments{
\item{name}{character}

\item{schemas}{list}
}
\value{
S3 Object
}
\description{
S3 Constructor for Class 'db_database'
}
\examples{
\dontrun{
test <- new_db_database(name)
}
}
