% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_proj_template.R
\name{new_db_proj_template}
\alias{new_db_proj_template}
\title{Constructor for S3 Class 'db_proj_template'}
\usage{
new_db_proj_template()
}
\value{
S3Class
}
\description{
Constructor for S3 Class 'db_proj_template'
}
\examples{
\dontrun{
new_db_proj_template()
}
}
