% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_class-db_schema.R
\name{new_db_schema}
\alias{new_db_schema}
\title{S3 Constructor for Class 'db_schema'}
\usage{
new_db_schema(x)
}
\arguments{
\item{x}{character}
}
\value{
S3 Object
}
\description{
S3 Constructor for Class 'db_schema'
}
\examples{
\dontrun{
test <- new_db_schema(raw_data_frame)
}
}
