% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_class-dbproj_char-tuple.R
\name{new_dbproj_char_tuple}
\alias{new_dbproj_char_tuple}
\title{S3 Constructor for Class 'dbproj_char_tuple'}
\usage{
new_dbproj_char_tuple(x)
}
\arguments{
\item{x}{character}
}
\value{
S3 Object
}
\description{
S3 Constructor for Class 'dbproj_char_tuple'
}
\examples{
\dontrun{
output <- new_dbproj_char_tuple(x)
}
}
