% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_class-dbproj_column.R
\name{new_dbproj_column}
\alias{new_dbproj_column}
\title{S3 Constructor for Class 'dbproj_column'}
\usage{
new_dbproj_column(name, type)
}
\arguments{
\item{name}{character}

\item{type}{character}
}
\value{
S3 Object
}
\description{
S3 Constructor for Class 'dbproj_column'
}
\examples{
\dontrun{
test <- new_dbproj_column(raw_data_frame)
}
}
