% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_class-dbproj_db.R
\name{new_dbproj_db}
\alias{new_dbproj_db}
\title{S3 Constructor for Class 'dbproj_db'}
\usage{
new_dbproj_db(name, schema)
}
\arguments{
\item{name}{character}

\item{schema}{list}
}
\value{
S3 Object
}
\description{
S3 Constructor for Class 'dbproj_db'
}
\examples{
\dontrun{
test <- new_dbproj_db(name)
}
}
