% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_class-dbproj_schema.R
\name{new_dbproj_schema}
\alias{new_dbproj_schema}
\title{S3 Constructor for Class 'dbproj_schema'}
\usage{
new_dbproj_schema(name, tables)
}
\arguments{
\item{name}{character}

\item{tables}{list}
}
\value{
S3 Object
}
\description{
S3 Constructor for Class 'dbproj_schema'
}
\examples{
\dontrun{
test <- new_dbproj_schema(raw_data_frame)
}
}
