% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_class-dbproj_table.R
\name{new_dbproj_table}
\alias{new_dbproj_table}
\title{S3 Constructor for Class 'dbproj_table'}
\usage{
new_dbproj_table(name, columns)
}
\arguments{
\item{name}{character}

\item{columns}{list}
}
\value{
S3 Object
}
\description{
S3 Constructor for Class 'dbproj_table'
}
\examples{
\dontrun{
test <- new_dbproj_table(raw_data_frame)
}
}
