% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-package_helper.R
\name{packageDoCall}
\alias{packageDoCall}
\title{Call Package Function with Parameters (and Validation)}
\usage{
packageDoCall(func_name, ...)
}
\arguments{
\item{func_name}{character - Name of Package Function}

\item{...}{R ellipsis - Arguments passed onto Package Function}
}
\value{
R Object
}
\description{
Call Package Function with Parameters (and Validation)
}
\examples{
\dontrun{
output <- packageDoCall('func', x = 3)
}
}
