% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-package_helper.R
\name{packageHandleDots}
\alias{packageHandleDots}
\title{Pre-Process and Validate R Ellipsis Dot Arguments}
\usage{
packageHandleDots(..., allow_na = FALSE, allow_dup = FALSE)
}
\arguments{
\item{...}{R ellipsis}

\item{allow_na}{logical - TRUE/FALSE}

\item{allow_dup}{logical - TRUE/FALSE}
}
\value{
list
}
\description{
Pre-Process and Validate R Ellipsis Dot Arguments
}
\examples{
\dontrun{
dot_args <- packageHandleDots(...)
}
}
