% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-s3_helper.R
\name{packageNameSpace}
\alias{packageNameSpace}
\title{List Content Names and Types for Package 'dbproj' Namespace}
\usage{
packageNameSpace()
}
\value{
data.frame
}
\description{
List Content Names and Types for Package 'dbproj' Namespace
}
\examples{
\dontrun{
output <- packageNameSpace()
}
}
