% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-package_helper.R
\name{packageS3Generics}
\alias{packageS3Generics}
\title{List S3 Generics for Package 'dbproj'}
\usage{
packageS3Generics()
}
\value{
character
}
\description{
List S3 Generics for Package 'dbproj'
}
\examples{
\dontrun{
output <- packageS3Generics()
}
}
