% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_class-dbproj_schema.R
\name{schema}
\alias{schema}
\title{S3 Helper Function for Class 'schema'}
\usage{
schema(name)
}
\arguments{
\item{name}{data.frame}
}
\value{
S3 Object
}
\description{
S3 Helper Function for Class 'schema'
}
\examples{
\dontrun{
test <- schema(IMF_DATA)
}
}
