% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_db_proj_template.R
\name{setup_db_proj_data}
\alias{setup_db_proj_data}
\title{Initialize Data for S3 Object of class 'db_proj_template'}
\usage{
setup_db_proj_data(obj, type)
}
\arguments{
\item{obj}{S3 Object}

\item{type}{character}
}
\description{
Initialize Data for S3 Object of class 'db_proj_template'
}
\examples{
\dontrun{
setup_db_proj_data(template, 'mssql')
}
}
