% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_db_proj_template.R
\name{setup_db_proj_metadata}
\alias{setup_db_proj_metadata}
\title{Initialize Metadata for S3 Object of class 'db_proj_template'}
\usage{
setup_db_proj_metadata(obj, type)
}
\arguments{
\item{obj}{S3 Object}

\item{type}{character}
}
\description{
Initialize Metadata for S3 Object of class 'db_proj_template'
}
\examples{
\dontrun{
setup_db_proj_metadata(template, 'mssql')
}
}
