% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_db_proj_template.R
\name{setup_db_proj_template}
\alias{setup_db_proj_template}
\title{Setup Internal Data for S3 Object of class 'db_proj_template'}
\usage{
setup_db_proj_template(obj)
}
\arguments{
\item{obj}{S3 Object}
}
\description{
Setup Internal Data for S3 Object of class 'db_proj_template'
}
\examples{
\dontrun{
setup_db_proj_template(template)
}
}
