% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_class-dbproj_table.R
\name{table}
\alias{table}
\title{S3 Helper Function for Class 'table'}
\usage{
table(name)
}
\arguments{
\item{name}{data.frame}
}
\value{
S3 Object
}
\description{
S3 Helper Function for Class 'table'
}
\examples{
\dontrun{
test <- table(IMF_DATA)
}
}
