% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_methods-db_database.R
\name{update.db_database}
\alias{update.db_database}
\title{S3 Method - CRUD 'Update' Operation}
\usage{
\method{update}{db_database}(obj, ...)
}
\arguments{
\item{obj}{S3 Object}

\item{...}{R ellipsis}
}
\description{
S3 Method - CRUD 'Update' Operation
}
\examples{
\dontrun{
update(obj, ...)
}
}
