% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_methods-db_schema.R
\name{update.db_schema}
\alias{update.db_schema}
\title{S3 Method - CRUD 'Update' Operation}
\usage{
\method{update}{db_schema}(obj, ...)
}
\arguments{
\item{obj}{S3 Object}

\item{...}{R ellipsis}
}
\description{
S3 Method - CRUD 'Update' Operation
}
\examples{
\dontrun{
update(obj, ...)
}
}
