% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_generics.R
\name{update}
\alias{update}
\title{S3 Generic - CRUD 'Update' Operation}
\usage{
update(obj, ...)
}
\arguments{
\item{obj}{S3 Object}

\item{...}{R ellipsis}
}
\description{
S3 Generic - CRUD 'Update' Operation
}
\examples{
\dontrun{
update(obj, ...)
}
}
