% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-s3_validate_base.R
\name{validate.character}
\alias{validate.character}
\title{S3 Validator for Class 'character'}
\usage{
\method{validate}{character}(obj, ...)
}
\arguments{
\item{obj}{S3 Object}

\item{...}{R ellipsis}
}
\value{
S3 Object
}
\description{
S3 Validator for Class 'character'
}
\examples{
\dontrun{
test <- validate.character(s3_obj)
}
}
