% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_class-dbproj_char-tuple.R
\name{validate.dbproj_char_tuple}
\alias{validate.dbproj_char_tuple}
\title{S3 Validator for Class 'dbproj_char_tuple'}
\usage{
\method{validate}{dbproj_char_tuple}(obj, ..., bool = FALSE)
}
\arguments{
\item{obj}{S3 Object}

\item{...}{R ellipsis}

\item{bool}{TRUE/FALSE}
}
\value{
S3 Object
}
\description{
S3 Validator for Class 'dbproj_char_tuple'
}
\examples{
\dontrun{
test <- validate.dbproj_char_tuple(s3_obj, bool = TRUE)
}
}
