% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_class-dbproj_column.R
\name{validate.dbproj_column}
\alias{validate.dbproj_column}
\title{S3 Validator for Class 'dbproj_column'}
\usage{
\method{validate}{dbproj_column}(obj, ..., bool = FALSE)
}
\arguments{
\item{obj}{S3 Object}

\item{...}{R ellipsis}

\item{bool}{TRUE/FALSE}
}
\value{
S3 Object
}
\description{
S3 Validator for Class 'dbproj_column'
}
\examples{
\dontrun{
test <- validate.dbproj_column(s3_obj, bool = TRUE)
}
}
