% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_class-dbproj_db.R
\name{validate.dbproj_db}
\alias{validate.dbproj_db}
\title{S3 Generic - Validate Format of 'dbproj_db'}
\usage{
\method{validate}{dbproj_db}(obj, ..., bool = FALSE)
}
\arguments{
\item{obj}{S3 Object}

\item{...}{R ellipsis}

\item{bool}{TRUE/FALSE}
}
\value{
S3 Object
}
\description{
S3 Generic - Validate Format of 'dbproj_db'
}
\examples{
\dontrun{
test <- validate.dbproj_db(s3_obj, bool = FALSE)
}
}
