% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_class-dbproj_table.R
\name{validate.dbproj_table}
\alias{validate.dbproj_table}
\title{S3 Validator for Class 'dbproj_table'}
\usage{
\method{validate}{dbproj_table}(obj, ..., bool = FALSE)
}
\arguments{
\item{obj}{S3 Object}

\item{...}{R ellipsis}

\item{bool}{TRUE/FALSE}
}
\value{
S3 Object
}
\description{
S3 Validator for Class 'dbproj_table'
}
\examples{
\dontrun{
test <- validate.dbproj_table(s3_obj, bool = TRUE)
}
}
