% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-s3_validate_base.R
\name{validate.integer}
\alias{validate.integer}
\title{S3 Validator for Class 'integer'}
\usage{
\method{validate}{integer}(obj, ...)
}
\arguments{
\item{obj}{S3 Object}

\item{...}{R ellipsis}
}
\value{
S3 Object
}
\description{
S3 Validator for Class 'integer'
}
\examples{
\dontrun{
test <- validate.integer(s3_obj)
}
}
