% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-s3_validate_base.R
\name{validate.list}
\alias{validate.list}
\title{S3 Validator for Class 'list'}
\usage{
\method{validate}{list}(obj, ..., bool = FALSE)
}
\arguments{
\item{obj}{S3 Object}

\item{...}{R ellipsis}

\item{bool}{TRUE/FALSE}
}
\value{
S3 Object
}
\description{
S3 Validator for Class 'list'
}
\examples{
\dontrun{
test <- validate.list(s3_obj, bool = TRUE)
}
}
