% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-s3_validate_base.R
\name{validate.logical}
\alias{validate.logical}
\title{S3 Validator for Class 'logical'}
\usage{
\method{validate}{logical}(obj, ..., bool = FALSE)
}
\arguments{
\item{obj}{S3 Object}

\item{...}{R ellipsis}

\item{bool}{TRUE/FALSE}
}
\value{
S3 Object
}
\description{
S3 Validator for Class 'logical'
}
\examples{
\dontrun{
test <- validate.logical(s3_obj, bool = TRUE)
}
}
