% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-s3_validation_methods.R
\name{validate.numeric}
\alias{validate.numeric}
\title{S3 Validator for Class 'numeric'}
\usage{
\method{validate}{numeric}(
  obj,
  allow_empty = FALSE,
  allow_na = FALSE,
  single = FALSE,
  check_names = FALSE,
  check_int = FALSE,
  check_neg = FALSE,
  ...
)
}
\arguments{
\item{obj}{S3 Object}

\item{allow_empty}{logical}

\item{allow_na}{logical}

\item{single}{logical}

\item{check_names}{logical}

\item{check_int}{logical}

\item{check_neg}{logical}

\item{...}{R ellipsis}
}
\value{
S3 Object
}
\description{
S3 Validator for Class 'numeric'
}
\examples{
\dontrun{
test <- validate.numeric(s3_obj)
}
}
