% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_generics.R
\name{validate}
\alias{validate}
\title{S3 Generic - Validate Format of S3 Object Instances}
\usage{
validate(obj, ...)
}
\arguments{
\item{obj}{S3 Object}

\item{...}{R ellipsis}
}
\value{
S3 Object
}
\description{
S3 Generic - Validate Format of S3 Object Instances
}
\examples{
\dontrun{
test <- validate(s3_obj)
}
}
