% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_class-db_database.R
\name{validate_db_database}
\alias{validate_db_database}
\title{S3 Validator for Class 'db_database'}
\usage{
validate_db_database(obj, bool)
}
\arguments{
\item{obj}{S3 Object}

\item{bool}{TRUE/FALSE}
}
\value{
S3 Object
}
\description{
S3 Validator for Class 'db_database'
}
\examples{
\dontrun{
test <- validate_db_database(s3_obj, FALSE)
}
}
