% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_proj_template.R
\name{validate_db_proj_template}
\alias{validate_db_proj_template}
\title{Validate that S3 Object is Properly Formatted}
\usage{
validate_db_proj_template(obj, bool)
}
\arguments{
\item{obj}{S3 Object - R Object to be validated against desired format}

\item{bool}{TRUE/FALSE - Optionally specify if input 'obj' or logical TRUE/FALSE}
}
\value{
R Object
}
\description{
Validate that S3 Object is Properly Formatted
}
\examples{
\dontrun{
validate_db_proj_template(obj, TRUE)
validate_db_proj_template(obj, FALSE)
}
}
