% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_class-db_schema.R
\name{validate_db_schema}
\alias{validate_db_schema}
\title{S3 Validator for Class 'db_schema'}
\usage{
validate_db_schema(obj, bool)
}
\arguments{
\item{obj}{S3 Object}

\item{bool}{TRUE/FALSE}
}
\value{
S3 Object
}
\description{
S3 Validator for Class 'db_schema'
}
\examples{
\dontrun{
test <- validate_db_schema(s3_obj, FALSE)
}
}
