% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_methods-dbproj_char-tuple.R
\name{equals.dbproj_char_tuple}
\alias{equals.dbproj_char_tuple}
\title{S3 Method - Test Equality of S3 'dbproj_char_tuple' Instances}
\usage{
\method{equals}{dbproj_char_tuple}(obj, x, ...)
}
\arguments{
\item{obj}{S3 Object}

\item{x}{S3 Object}

\item{...}{R ellipsis}
}
\value{
logical
}
\description{
S3 Method - Test Equality of S3 'dbproj_char_tuple' Instances
}
\examples{
\dontrun{
test <- equals(s3_obj, s3_obj)
}
}
