% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_methods-dbproj_char-tuple.R
\name{insert.dbproj_char_tuple}
\alias{insert.dbproj_char_tuple}
\title{S3 Method - CRUD 'Insert' Operation}
\usage{
\method{insert}{dbproj_char_tuple}(obj, x, in_place = TRUE, ...)
}
\arguments{
\item{obj}{S3 Object}

\item{x}{character}

\item{in_place}{logical}

\item{...}{R ellipsis}
}
\description{
S3 Method - CRUD 'Insert' Operation
}
\examples{
\dontrun{
insert(obj, x)
output <- insert(obj, x, in_place = FALSE)
}
}
