% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-s3_validation_methods.R
\name{validate.logical}
\alias{validate.logical}
\title{S3 Validator for Class 'logical'}
\usage{
\method{validate}{logical}(
  obj,
  allow_empty = FALSE,
  allow_na = FALSE,
  single = FALSE,
  check_names = FALSE,
  ...
)
}
\arguments{
\item{obj}{S3 Object}

\item{allow_empty}{logical}

\item{allow_na}{logical}

\item{single}{logical}

\item{check_names}{logical}

\item{...}{R ellipsis}
}
\value{
S3 Object
}
\description{
S3 Validator for Class 'logical'
}
\examples{
\dontrun{
test <- validate.logical(s3_obj)
}
}
