% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-s3_validate_base.R
\name{validate.numeric}
\alias{validate.numeric}
\title{S3 Validator for Class 'numeric'}
\usage{
\method{validate}{numeric}(obj, ...)
}
\arguments{
\item{obj}{S3 Object}

\item{...}{R ellipsis}
}
\value{
S3 Object
}
\description{
S3 Validator for Class 'numeric'
}
\examples{
\dontrun{
test <- validate.numeric(s3_obj)
}
}
